Sanitize cursor_col in move_backward
If the user moved the cursor using free movement (h/j/k/l) to a position
beyond input on a line with input, and then used back-word, the jump
would be off by the amount of characters that would be in the space
between the cursor and the input.

Before calculating the start of the word, snap cursor_col to the
last character of input.