Add smart_copy option
This has been inspired from the Elementary OS terminal. This option
allows the user to copy and paste naturally just like in any other
application with minimal feature loss. And, more importantly, reduces
the habit where I press CTRL + SHIFT + C in my browser and it opens
the devtools.

I needed to change the copy behavior so it deselected any selected text
after copying. This allows the CTRL + C, CTRL + V combo to work without
having to click on the terminal in between.