Add option to make font size follow monitor DPI
Add a new "dpi_aware" option which, when enabled, makes Termite try to
adjust the absolute font size to match the expected physical size in
points for the current monitor. This causes text to have the same
size regardless of which monitor the window is displayed onto. The
adjustment is updated live as the window changes monitors as well.