Implement copy_mouse_selection option
Add a new "copy_mouse_selection" option. When enabled, selecting text
with the mouse will also copy it to both the CLIPBOARD and PRIMARY
selections. When disabled, selected text is only copied to the PRIMARY
one. Note that this does not change the behaviour in command and
selection modes. In particular, while in selection mode the "y" (yank)
key binding was already available to send the selection to CLIPBOARD.

Fixes #16